refactor(compare): aplica os 8 achados da revisão do PR #359
Confirmados:
- ExistingVerdict deixa de ser re-declarado: ComparisonPanel e
  DivergenceActionsPanel importam VerdictInfo de lib/compare-reviews
  (origem real do valor).
- UnansweredResponse vira Pick<FieldResponse, ...> de compare-types —
  drift do tipo canônico agora é erro de compilação.
- onVerdict do DivergenceActionsPanel estreitado para
  (verdict: string) => void: o chosenResponseId era parâmetro morto
  (voto é papel do AgreementGroup); o handler mais largo do pai segue
  atribuível por contravariância.
- useMemo removido do UnansweredNotice: fieldResponses muda de
  identidade a cada render do ComparePage, então o memo recomputava
  exatamente quando importaria; alinha com no-manual-memoization.

Plausíveis:
- formatVerdictDisplay (variante join ", ", sem traduzir marcadores)
  hoistado para lib/verdict-display, consumido por
  DivergenceActionsPanel e LlmErrorCard; documenta a distinção
  intencional frente a lib/reviews/verdict-format.
- respondentKey() exportado de compare-types: a chave de dedup
  respondent_id ?? id, antes acordada só por comentário entre
  page.tsx (gate minHumans) e UnansweredNotice, vira estrutural.
- SuggestFieldDialog + suggestOpen colocados no DivergenceActionsPanel
  junto do único gatilho (botão Sugerir); o pai passa fields no lugar
  de onSuggest. Aresta conhecida: navegar para campo não-divergente
  com o dialog aberto agora o fecha (antes remontava via key).
- isMulti declarado como boolean (!!fieldOptions?.length) elimina os
  dois casts !! nos call sites; narrowing de fieldOptions preservado.

Validação: tsc sem erros novos (18 pré-existentes dos fixtures de
ComparePage), react-doctor 100/100 em compare/, vitest 863/863.
